+++ node-api/src/main/kotlin/net/corda/nodeapi/internal/serialization/ClientContexts.kt
@@ -1,0 +1,15 @@
+@file:JvmName("ClientContexts")
+package net.corda.nodeapi.internal.serialization
+
+import net.corda.core.serialization.SerializationContext
+import net.corda.core.serialization.SerializationDefaults
+
+/*
+ * Serialisation contexts for the client.
+ */
+val KRYO_RPC_CLIENT_CONTEXT = SerializationContextImpl(KryoHeaderV0_1,
+        SerializationDefaults.javaClass.classLoader,
+        GlobalTransientClassWhiteList(BuiltInExceptionsWhitelist()),
+        emptyMap(),
+        true,
+        SerializationContext.UseCase.RPCClient)
